game Menu Commit
casinoGame.java is my best implementation into what we need for the game menu.
mainMenu1 is now what casinoGame.java is however with a main so you can run it and see what it looks like.
Let me know if there is something else i should work on through groupme